requirements.txt
beautifulsoup4
flask
APScheduler
mysql-connector-python
